+++ README.md
@@ -1,8 +1,10 @@
 # Usage
 
-**Please add your server IP and domain to the whitelist in the Wechat management system at first, and keep the domain consistent with the whitelist defined in the `package.json`.**
+**Please add your server IP and domain to the whitelist in the Wechat Official Accounts Platform at first.**
 
 You should create the `appid` and `appsecret` file in the root path before start.
+
+## wxShare
 
 ```html
 <script>
@@ -27,3 +29,13 @@ window.wxShare.setShareData({
 })
 </script>
 ```
+
+## wxAuth
+
+**Please set the `authPostDataURIDomainWhitelist` in package.json at first**
+
+Redirect to `{serverPath}/wxAuth?redirect_uri={}&postdata_uri={}&error_uri={}`, server will redirect user to wechat auth page.
+
+After auth successful, server will post user information to the `postdata_uri`, then user will be redirected to the `redirect_uri`.
+
+If there has any error, server will redirect user to `error_uri`.
