+++ package.json
@@ -37,8 +37,7 @@
     "node": ">=8.10.0"
   },
   "authPostDataURIDomainWhitelist": [
-    "hduzplus.xyz",
-    "baidu.com"
+    "hduzplus.xyz"
   ],
   "wechatJSURL": "http://res2.wx.qq.com/open/js/jweixin-1.6.0.js",
   "serverURL": "https://static.hduzplus.xyz"
