+++ plugins/auth-domain-interceptor.js
@@ -1,4 +1,5 @@
 const { URL } = require('url');
+const { IS_DEV } = require('../constants');
 const { authPostDataURIDomainWhitelist } = require('../package.json');
 
 const whitelist = new Set(authPostDataURIDomainWhitelist);
@@ -8,9 +9,9 @@ module.exports = async function authPostDataURIDomainInterceptor (ctx, next) {
     postdata_uri: postdataURI,
   } = ctx.request.query;
 
-  const { host: postdataHost } = new URL(postdataURI);
+  const { host: postdataHost } = new URL(decodeURIComponent(postdataURI));
 
-  if (whitelist.has(postdataHost)) {
+  if (IS_DEV || whitelist.has(postdataHost)) {
     await next();
     return;
   }
