+++ plugins/error-handler.js
@@ -1,9 +1,10 @@
 const log = require('../utils/log');
 
-module.exports = async function errorHandler (_, next) {
+module.exports = async function errorHandler (ctx, next) {
   try {
     await next();
   } catch (e) {
+    ctx.status = 400;
     log(`[Error] ${e.message}`);
   }
 };
